requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
great_expectations==0.18.8
pytest==8.0.0
pytest-spark==0.7.0
